feat(simulation): add car km and type question component
