+++ packages/mobile/src/adapters/simulation-results/store/simulation-store.ts
@@ -1,4 +1,4 @@
-import { CarbonFootprintDto, SimulationDto } from 'carbon-cut-commons';
+import { CarbonFootprintDto, CarSize, FuelType, SimulationDto } from 'carbon-cut-commons';
 import { injectable } from 'inversify';
 import { AnswerToSave, SimulationStore } from '../../../domain/ports/stores/simulation-store';
 import { navigate, Routes } from '../../../infrastructure/root-navigation';
@@ -14,7 +14,14 @@ export class ReduxSimulationStore implements SimulationStore {
   }
 
   getSimulationsAnswers() {
-    return selectSimulationAnswers();
+    const answers = selectSimulationAnswers();
+    const transport = answers?.transport
+      ? {
+          ...answers?.transport,
+          car: { ...answers.transport.car, fuelType: FuelType.diesel, fuelConsumption: 6.5, carSize: CarSize.medium },
+        }
+      : { car: { fuelType: FuelType.diesel, carSize: CarSize.medium } };
+    return { ...answers, transport } as SimulationDto;
   }
 
   saveAnswer(sector: keyof SimulationDto, answer: Partial<AnswerToSave>) {
