+++ packages/mobile/src/infrastructure/simulation/pages/alimentation/meals.tsx
@@ -1,40 +1,39 @@
+import { NavigationProp } from '@react-navigation/native';
 import { Text, useTheme } from '@rneui/themed';
 import { MealsAnswer } from 'carbon-cut-commons';
 import { useEffect, useState } from 'react';
 import { ScrollView, StyleProp, StyleSheet, View, ViewStyle } from 'react-native';
-import { useSelector } from 'react-redux';
-import { selectIsLoading } from '../../../adapters/commons/store/selectors/loading-selectors';
-import { MealViewModel, WebMealsQuestionPresenter } from '../../../adapters/simulation/presenters/web-meals-question.presenter';
-import { AnswerValidator } from '../../../domain/entites/answer-validator';
-import { MealsQuestionPresenterToken } from '../../../domain/ports/presenters/question.presenter';
-
 import {
-  CarbonFootprintSimulationUseCase,
-  CarbonFootprintSimulationUseCaseToken,
-} from '../../../domain/usecases/carbon-footprint-simulation.usescase';
-import { SaveSimulationAnswerUseCase, SaveSimulationAnswerUseCaseToken } from '../../../domain/usecases/save-simulation-answer.usecase';
-import { SetInputAnswerUseCase, SetInputAnswerUseCaseToken } from '../../../domain/usecases/set-input-answer.usecase';
-import { errorStyle } from '../../app/style';
-import { diContainer } from '../../inversify.config';
-import InputAnswers from '../components/input-answers';
-import Question from '../components/question';
-import SubmitButton from '../components/submit-button';
+  MealViewModel,
+  WebMealsQuestionPresenter,
+} from '../../../../adapters/simulation/presenters/alimentation/web-meals-question.presenter';
+import { AnswerValidator } from '../../../../domain/entites/answer-validator';
+import { MealsQuestionPresenterToken } from '../../../../domain/ports/presenters/question.presenter';
+import { SaveSimulationAnswerUseCase, SaveSimulationAnswerUseCaseToken } from '../../../../domain/usecases/save-simulation-answer.usecase';
+import { SetInputAnswerUseCase, SetInputAnswerUseCaseToken } from '../../../../domain/usecases/set-input-answer.usecase';
+import { errorStyle } from '../../../app/style';
+import { diContainer } from '../../../inversify.config';
+import { RootStackParamList, Routes } from '../../../root-navigation';
+import InputAnswers from '../../components/input-answers';
+import Question from '../../components/question';
+import SubmitButton from '../../components/submit-button';
 
 export type MealsAnswerKeys = keyof MealsAnswer;
+type MealsNavigationProp = NavigationProp<RootStackParamList, Routes.Meals>;
+type MealsProps = {
+  navigation: MealsNavigationProp;
+  containerStyle: StyleProp<ViewStyle>;
+};
 
-export default function Meals({ containerStyle }: { containerStyle: StyleProp<ViewStyle> }) {
+export default function Meals({ navigation, containerStyle }: MealsProps) {
   const { theme } = useTheme();
   const [presenter] = useState<WebMealsQuestionPresenter>(diContainer.get<WebMealsQuestionPresenter>(MealsQuestionPresenterToken));
   const [setInputAnswerUseCase] = useState<SetInputAnswerUseCase>(diContainer.get<SetInputAnswerUseCase>(SetInputAnswerUseCaseToken));
   const [saveSimulationAnswerUseCase] = useState<SaveSimulationAnswerUseCase>(
     diContainer.get<SaveSimulationAnswerUseCase>(SaveSimulationAnswerUseCaseToken),
   );
-  const [carbonFootprintSimulationUseCase] = useState<CarbonFootprintSimulationUseCase>(
-    diContainer.get<CarbonFootprintSimulationUseCase>(CarbonFootprintSimulationUseCaseToken),
-  );
 
   const [viewModel, updateViewModel] = useState<MealViewModel>(presenter.viewModel);
-  const isLoading = useSelector(selectIsLoading);
 
   useEffect(() => {
     presenter.onViewModelChanges(updateViewModel);
@@ -53,7 +52,7 @@ export default function Meals({ containerStyle }: { containerStyle: StyleProp<Vi
 
   const runCalculation = (): void => {
     saveSimulationAnswerUseCase.execute({ sector: 'alimentation', answerKey: 'meals', answer: presenter.answerValues });
-    carbonFootprintSimulationUseCase.execute();
+    navigation.navigate(Routes.CarKmType, { containerStyle });
   };
 
   return (
@@ -64,12 +63,7 @@ export default function Meals({ containerStyle }: { containerStyle: StyleProp<Vi
         </Question>
       </ScrollView>
       {viewModel.formError ? <Text style={errorStyle(theme).error}>{viewModel.formError}</Text> : null}
-      <SubmitButton
-        isLastQuestion={true}
-        canSubmit={viewModel.canSubmit}
-        isLoading={isLoading}
-        nextButtonClicked={() => runCalculation()}
-      />
+      <SubmitButton isLastQuestion={false} canSubmit={viewModel.canSubmit} nextButtonClicked={() => runCalculation()} />
     </View>
   );
 }
