+++ packages/mobile/src/infrastructure/simulation/pages/alimentation/milk-type.tsx
@@ -5,14 +5,14 @@ import {
   MilkTypeAnswer,
   MilkTypeViewModel,
   WebMilkTypeQuestionPresenter,
-} from '../../../adapters/simulation/presenters/web-milk-type-question.presenter';
-import { MilkTypeQuestionPresenterToken } from '../../../domain/ports/presenters/question.presenter';
-import { SaveSimulationAnswerUseCase, SaveSimulationAnswerUseCaseToken } from '../../../domain/usecases/save-simulation-answer.usecase';
-import { diContainer } from '../../inversify.config';
-import { RootStackParamList, Routes } from '../../root-navigation';
-import Question from '../components/question';
-import SelectableAnswers from '../components/selectable-answers';
-import SubmitButton from '../components/submit-button';
+} from '../../../../adapters/simulation/presenters/alimentation/web-milk-type-question.presenter';
+import { MilkTypeQuestionPresenterToken } from '../../../../domain/ports/presenters/question.presenter';
+import { SaveSimulationAnswerUseCase, SaveSimulationAnswerUseCaseToken } from '../../../../domain/usecases/save-simulation-answer.usecase';
+import { diContainer } from '../../../inversify.config';
+import { RootStackParamList, Routes } from '../../../root-navigation';
+import Question from '../../components/question';
+import SelectableAnswers from '../../components/selectable-answers';
+import SubmitButton from '../../components/submit-button';
 
 type MilkTypeNavigationProp = NavigationProp<RootStackParamList, Routes.MilkType>;
 type MilkTypeProps = {
@@ -32,7 +32,7 @@ export default function MilkType({ navigation, containerStyle }: MilkTypeProps) 
   }, [presenter, presenter.viewModel]);
 
   const setSelectedMilkType = (answer: MilkTypeAnswer): void => {
-    presenter.setAnswer(answer.value);
+    presenter.setSelectedAnswer(answer.value);
   };
 
   const saveAnswer = (): void => {
