+++ packages/mobile/src/tests/answer.ts
@@ -1,0 +1,9 @@
+export const unselectedAnswerStyle = {
+  backgroundColor: 'transparent',
+  borderColor: '#2d8456',
+};
+
+export const selectedAnswerStyle = {
+  backgroundColor: '#2d8456',
+  borderColor: '#2d8456',
+};
